+++ src/TechDivision/ApplicationServer/Interfaces/ClassLoaderInterface.php
@@ -44,12 +44,12 @@ interface ClassLoaderInterface
     public function register($throw = true, $prepend = false);
 
     /**
-     * Factory method that adds a initialized class loader to the passed application.
+     * Visitor method that adds a initialized class loader to the passed application.
      *
      * @param \TechDivision\Application\Interfaces\ApplicationInterface         $application   The application instance
      * @param \TechDivision\ApplicationServer\Api\Node\ClassLoaderNodeInterface $configuration The class loader configuration node
      *
      * @return void
      */
-    public static function get(ApplicationInterface $application, ClassLoaderNodeInterface $configuration = null);
+    public static function visit(ApplicationInterface $application, ClassLoaderNodeInterface $configuration = null);
 }