+++ src/TechDivision/ApplicationServer/SplClassLoader.php
@@ -60,14 +60,14 @@ class SplClassLoader extends \Stackable
     const CLASS_MAP = 'SplClassLoader.classMap';
 
     /**
-     * Factory method that adds a initialized class loader to the passed application.
+     * Visitor method that adds a initialized class loader to the passed application.
      *
      * @param \TechDivision\Application\Interfaces\ApplicationInterface         $application   The application instance
      * @param \TechDivision\ApplicationServer\Api\Node\ClassLoaderNodeInterface $configuration The class loader configuration node
      *
      * @return void
      */
-    public static function get(ApplicationInterface $application, ClassLoaderNodeInterface $configuration = null)
+    public static function visit(ApplicationInterface $application, ClassLoaderNodeInterface $configuration = null)
     {
 
         // load the web application path we want to register the class loader for